Update Dockerfile and workflow for standardized deployment (#21)
* Update Dockerfile and workflow for standardized deployment

Replaced exposed ports with a single port 80 in the Dockerfile for consistency and updated the entry point to explicitly bind to it. Adjusted GitHub Actions workflow to use a capitalized environment name for improved naming convention.

* Update Docker ENTRYPOINT to use wildcard host

Changed the ENTRYPOINT URL from `http://0.0.0.0:80` to `http://+`. This ensures compatibility with systems that rely on the wildcard host for binding to all network interfaces.

* Update Dockerfile to specify IPv6-compatible URL binding

Changed the ENTRYPOINT to bind the application to IPv6-compatible addresses by updating the --urls argument. This ensures better compatibility and support for systems using IPv6 alongside IPv4.

* Update Dockerfile to configure ASP.NET Core URLs via ENV

Replaced hardcoded URLs in the ENTRYPOINT with an environment variable `ASPNETCORE_URLS`. This improves flexibility and allows easier reconfiguration without modifying the Dockerfile.