Fix issue with Migration Update migrate.yml (#33)
* Fix issue with Migration Update migrate.yml

* Reorder steps in test workflow for logical execution

Moved the "Checkout code" step above the "Anchore Scan" step to ensure the repository is properly checked out before running the scan. This fixes the sequence for a more consistent and reliable workflow.

* Update CI/CD workflows for clarity and consistency

Renamed job and step names in workflows to improve readability and better reflect their roles. Adjusted Anchore scan configuration for compatibility and updated terminology to align with conventions. These changes enhance maintainability and ensure consistent practices across workflows.

* Refine CI workflow for tagging and release conditions

Updated the CI workflow to ensure tags are only applied when pushing to the main branch and prevent duplicate version tags. This change improves release tagging logic and aligns conditions with expected behavior.

* Update version to 0.0.11 in eKIBRA.info

Bumped the version from 0.0.10 to 0.0.11 to reflect changes or improvements in the software. This ensures proper version tracking and alignment with future updates.

* Remove Anchore scan from tests, add image scan to build workflow

The Anchore scan for codebase files was removed from the test workflow to streamline the process. Instead, a container image scan was added to the build workflow for better alignment with security practices focused on built artifacts. This ensures a more relevant and efficient vulnerability scanning process.