Task#37 - [Pipeline] - Change Dockerfile to be compatible with Azure env (#41)
* Update exposed ports and environment variables in Dockerfiles

Changed exposed ports to include 80/tcp and 443/tcp for HTTP/HTTPS standards. Updated `ASPNETCORE_HTTP_PORTS` and `ASPNETCORE_HTTPS_PORTS` to reflect the new mappings for better compatibility and flexibility.

* Update eKIBRA version to 0.0.16

Bump the version number from 0.0.15 to 0.0.16 in preparation for the latest release. This ensures version control alignment for new updates and improvements.

* Update exposed ports and environment variables in Dockerfile

Removed unnecessary ports 8080 and 8081 to simplify configuration. Adjusted related environment variables and health check to align with the updated port setup. This change ensures a cleaner and more secure configuration.

* Update HEALTHCHECK URL in Dockerfile

Updated the HEALTHCHECK command to remove the explicit port reference from the URL. This ensures compatibility with configurations where the application is bound to the default port.